working cut. Seperate out kafka consumer and producer from ES.
Added new project kafka-persistence. Unit tests are pending.
