+++ core/src/main/java/com/netflix/conductor/dao/KafkaProducerDAO.java
@@ -15,6 +15,14 @@
  */
 package com.netflix.conductor.dao;
 
+import com.netflix.conductor.common.metadata.events.EventExecution;
+import com.netflix.conductor.common.metadata.tasks.Task;
+import com.netflix.conductor.common.metadata.tasks.TaskExecLog;
+import com.netflix.conductor.common.run.Workflow;
+import com.netflix.conductor.core.events.queue.Message;
+
+import java.util.List;
+
 /**
  *
  * @author Manan
@@ -23,4 +31,14 @@ package com.netflix.conductor.dao;
 public interface KafkaProducerDAO {
 
     void send(String t, Object value);
+
+    void produceWorkflow(Workflow workflow);
+
+    void produceTask(Task task);
+
+    void produceTaskExecutionLogs(List<TaskExecLog> taskExecLogs);
+
+    void produceMessage(String queue, Message message);
+
+    void produceEventExecution(EventExecution eventExecution);
 }
