+++ es5-persistence/src/test/java/com/netflix/conductor/dao/es5/index/TestElasticSearchDAOV5.java
@@ -10,6 +10,7 @@ import com.netflix.conductor.common.run.SearchResult;
 import com.netflix.conductor.common.run.Workflow;
 import com.netflix.conductor.core.events.queue.Message;
 import com.netflix.conductor.dao.es5.index.query.parser.Expression;
+import com.netflix.conductor.dao.kafka.index.KafkaProducer;
 import com.netflix.conductor.elasticsearch.ElasticSearchConfiguration;
 import com.netflix.conductor.elasticsearch.ElasticSearchTransportClientProvider;
 import com.netflix.conductor.elasticsearch.EmbeddedElasticSearch;
@@ -37,6 +38,7 @@ import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import org.mockito.Mockito;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -101,7 +103,8 @@ public class TestElasticSearchDAOV5 {
 				.get();
 
 		ObjectMapper objectMapper = new ObjectMapper();
-		indexDAO = new ElasticSearchDAOV5(elasticSearchClient, configuration, objectMapper);
+		KafkaProducer kafkaProducer = Mockito.mock(KafkaProducer.class);
+		indexDAO = new ElasticSearchDAOV5(elasticSearchClient, configuration, objectMapper, kafkaProducer);
 	}
 
 	@AfterClass
