+++ kafka-persistence/build.gradle
@@ -1,0 +1,16 @@
+dependencies {
+    compile project(':conductor-core')
+    compile project(':conductor-es5-persistence')
+
+    compile "commons-io:commons-io:${revCommonsIo}"
+
+    compile "org.apache.kafka:kafka-clients:2.2.0"
+
+    compile "org.elasticsearch:elasticsearch:${revElasticSearch5}"
+    compile "org.elasticsearch.client:transport:${revElasticSearch5}"
+    compile "org.elasticsearch.client:elasticsearch-rest-client:${revElasticSearch5}"
+    compile "org.elasticsearch.client:elasticsearch-rest-high-level-client:${revElasticSearch5}"
+
+    testCompile "org.slf4j:slf4j-log4j12:${revSlf4jlog4j}"
+    testCompile "org.awaitility:awaitility:${revAwaitility}"
+}
