+++ kafka-persistence/src/main/resources/mappings_docType_task.json
@@ -1,0 +1,68 @@
+{
+  "task": {
+    "properties": {
+      "correlationId": {
+        "type": "keyword",
+        "index": true
+      },
+      "endTime": {
+        "type": "date",
+        "format": "strict_date_optional_time||epoch_millis"
+      },
+      "executionTime": {
+        "type": "long"
+      },
+      "input": {
+        "type": "text",
+        "index": true
+      },
+      "output": {
+        "type": "text",
+        "index": true
+      },
+      "queueWaitTime": {
+        "type": "long"
+      },
+      "reasonForIncompletion": {
+        "type": "keyword",
+        "index": true
+      },
+      "scheduledTime": {
+        "type": "date",
+        "format": "strict_date_optional_time||epoch_millis"
+      },
+      "startTime": {
+        "type": "date",
+        "format": "strict_date_optional_time||epoch_millis"
+      },
+      "status": {
+        "type": "keyword",
+        "index": true
+      },
+      "taskDefName": {
+        "type": "keyword",
+        "index": true
+      },
+      "taskId": {
+        "type": "keyword",
+        "index": true
+      },
+      "taskType": {
+        "type": "keyword",
+        "index": true
+      },
+      "updateTime": {
+        "type": "date",
+        "format": "strict_date_optional_time||epoch_millis"
+      },
+      "workflowId": {
+        "type": "keyword",
+        "index": true
+      },
+      "workflowType": {
+        "type": "keyword",
+        "index": true
+      }
+    }
+  }
+}
