+++ kafka-persistence/src/main/resources/template_tasklog.json
@@ -1,0 +1,85 @@
+{
+    "order": 0,
+    "template": "*task*log*",
+    "settings": {
+        "index": {
+            "refresh_interval": "1s"
+        }
+    },
+    "mappings": {
+        "event": {
+            "properties": {
+                "action": {
+                    "type": "keyword",
+                    "index": true
+                },
+                "created": {
+                    "type": "long"
+                },
+                "event": {
+                    "type": "keyword",
+                    "index": true
+                },
+                "id": {
+                    "type": "keyword",
+                    "index": true
+                },
+                "messageId": {
+                    "type": "keyword",
+                    "index": true
+                },
+                "name": {
+                    "type": "keyword",
+                    "index": true
+                },
+                "output": {
+                    "properties": {
+                        "workflowId": {
+                            "type": "keyword",
+                            "index": true
+                        }
+                    }
+                },
+                "status": {
+                    "type": "keyword",
+                    "index": true
+                }
+            }
+        },
+        "task_log": {
+            "properties": {
+                "createdTime": {
+                    "type": "long"
+                },
+                "log": {
+                    "type": "keyword",
+                    "index": true
+                },
+                "taskId": {
+                    "type": "keyword",
+                    "index": true
+                }
+            }
+        },
+        "message": {
+            "properties": {
+                "created": {
+                    "type": "long"
+                },
+                "messageId": {
+                    "type": "keyword",
+                    "index": true
+                },
+                "payload": {
+                    "type": "keyword",
+                    "index": true
+                },
+                "queue": {
+                    "type": "keyword",
+                    "index": true
+                }
+            }
+        }
+    },
+    "aliases": {}
+}
