+++ server/build.gradle
@@ -27,6 +27,7 @@ dependencies {
     compile project(':conductor-cassandra-persistence')
     compile project(':conductor-contribs')
     compile project(':conductor-es5-persistence')
+    compile project(':conductor-kafka-persistence')
     compile project(':conductor-grpc-server')
 
     compile "com.netflix.runtime:health-guice:${revHealth}"
